pyproject.toml
[project]
name = "hrm"
version = "0.1.0"
description = ""
authors = [
    {name = "jakhonma",email = "jahongirjorayev9503@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-phonenumber-field[phonenumberslite] (>=8.1.0,<9.0.0)",
    "django-phonenumber-field[phonenumbers] (>=8.1.0,<9.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
